+++ lib/lark_custom_bot/cards.ex
@@ -1,0 +1,5 @@
+defmodule LarkCustomBot.Cards do
+  def new(_card) do
+    {:ok, %{}}
+  end
+end